+++ client/package.json
@@ -14,6 +14,8 @@
     "caniuse-lite": "^1.0.30001252",
     "centrifuge": "^2.8.0",
     "cross-env": "^7.0.3",
+    "date-fns": "^2.24.0",
+    "formik": "^2.2.9",
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -23,7 +25,8 @@
     "styled-components": "^5.3.1",
     "web-vitals": "^1.0.1",
     "ws": "^8.2.2",
-    "yarn": "^1.22.11"
+    "yarn": "^1.22.11",
+    "yup": "^0.32.9"
   },
   "scripts": {
     "start": "cross-env PORT=3006 react-scripts start",
