+++ client/src/components/create_goal/create-edit-goal.utils.js
@@ -1,0 +1,9 @@
+import { parse, isDate } from 'date-fns';
+
+const parseDateString = (value, originalValue) => {
+  const parsedDate = isDate(originalValue) ? originalValue : parse(originalValue, 'yyyy-MM-dd', new Date());
+
+  return parsedDate;
+};
+
+export default parseDateString;
