+++ client/src/components/create_goal/icon/Star 1.png
