+++ client/src/components/create_goal/icon/Star 17.png
