+++ client/src/components/create_goal/icon/active.png
