+++ client/src/components/create_goal/icon/default.png
