+++ client/src/pages/Home.jsx
@@ -2,13 +2,13 @@
 import styled from 'styled-components';
 import MainNav from 'components/MainNav/MainNav';
 import SnackBar from 'components/snackbar/Snackbar';
-import SimpleModal from '../components/createGoal/CreateGoal';
+import SimpleModal from '../components/create_goal/CreateGoal';
 import Deletemodal from '../components/Deletemodal/Deletemodal';
 import DeleteSuccessModal from '../components/DeleteSuccess/DeleteSuccessModal';
 import BasicModal from '../components/EditGoal/EditGoal';
 import MissionVisionContainer from '../components/header/Header';
 import Mainside from '../components/mainside/Mainside';
-//import EditMission from '../components/modal/EditMission';
+// import EditMission from '../components/modal/EditMission';
 import OrganizationVisionEditModal from '../components/organization_vision/org_edit_vision/modal/EditOrgVisionModal';
 
 function App() {
