+++ client/src/redux/store.js
@@ -2,11 +2,12 @@ import { configureStore } from '@reduxjs/toolkit';
 import deleteGoalReducer from './deleteGoal.slice';
 import editMissionReducer from './editMission.slice';
 import newGoalReducer from './newGoalSlice';
+import createAndEditGoalReducer from './organizationGoal.slice';
 import visionReducer from './organizationVision.slice';
+import showGoalSliceReducer from './showGoalSlice';
 import snackbarReducer from './snackbar.slice';
 import createGoalModalReducer from './toggleCreateGoalModal.slice';
 import editGoalModalReducer from './toggleEditGoalModal.slice';
-import showGoalSliceReducer from './showGoalSlice';
 
 const store = configureStore({
   reducer: {
@@ -18,6 +19,7 @@ const store = configureStore({
     newGoal: newGoalReducer,
     snackbar: snackbarReducer,
     goals: showGoalSliceReducer,
+    organizationCreateAndEditGoal: createAndEditGoalReducer,
   },
 });
